Added a small shop
Apologies, this should have been on a new shop branch!

Guide: a controller which generates buttons based on items given in the inspector, interface used for customer which is linked to the player (in this case) and checks to see if their is enough money (and other checks) before purchasing items which are controlled using other controllers in game.

Lessons: add branches for each feature! Please!