Skill Manager and Menu
Added a way to unlock skills which require either prior skills or a set amount of skill points.

Skills are unlocked per level up.

Added a simple menu which was manually created (with prefabs for the buttons) which are used to set specific skills.

Added a rounded corner sprite which makes the UI a bit cleaner.

To Do:
add a 'menu controller' to show the skill menu
fancy up the skill menu so skills unlocked are greyed out and then highlighted when unlocked
add an 'animation controller' which handles levelling up, skill unlocks etc.